Address review: reorganize file tree structure
